+++ src/BackupHandlers/BackupHandlerInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Spatie\Backup\BackupHandlers;
+
+interface BackupHandlerInterface {
+
+    /**
+     * Returns an array of files which should be backed up.
+     *
+     * @return array
+     */
+    public function getFilesToBeBackedUp();
+}