+++ src/BackupServiceProvider.php
@@ -1,10 +1,11 @@
-<?php namespace Spatie\Backup;
+<?php
+
+namespace Spatie\Backup;
 
 use Illuminate\Support\ServiceProvider;
 
 class BackupServiceProvider extends ServiceProvider
 {
-
     /**
      * Indicates if loading of the provider is deferred.
      *
@@ -20,7 +21,7 @@ class BackupServiceProvider extends ServiceProvider
     public function boot()
     {
         $this->publishes([
-            __DIR__.'/Assets/config/laravel-backup.php' => config_path('laravel-backup.php'),
+            __DIR__.'/config/laravel-backup.php' => config_path('laravel-backup.php'),
         ]);
     }
 
@@ -31,15 +32,13 @@ class BackupServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $databaseBuilder = new DatabaseBuilder();
-
-        $this->app['command.db:backup'] = $this->app->share(
-            function ($app) use ($databaseBuilder) {
-                return new Commands\BackupCommand($databaseBuilder);
+        $this->app['command.backup:run'] = $this->app->share(
+            function ($app)  {
+                return new Commands\BackupCommand();
             }
         );
 
-        $this->commands('command.db:backup');
+        $this->commands('command.backup:run');
     }
 
     /**
@@ -49,6 +48,6 @@ class BackupServiceProvider extends ServiceProvider
      */
     public function provides()
     {
-        return ['command.db:backup'];
+        return ['command.backup:run'];
     }
 }