+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `laravel-backup` will be documented in this file
 
+### 2.0.0
+- Added support to backup directories and individual files
+- Configuration file changed
+- Refactored all classes
+
 ### 1.2.0
 - Added support to backup to multiple filesystems at once